add SolaceJavaAutoConfigurationBase
